src/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-auth-ldap==1.7.0
django-filter==2.1.0
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-ldap==3.2.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
